# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  game_eating_pizza_internal_api_dto.DungeonResponse:
    properties:
      created_at:
        type: string
      difficulty:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      start_time:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  game_eating_pizza_internal_api_dto.PlayerResponse:
    properties:
      created_at:
        type: string
      experience:
        type: integer
      gold:
        type: integer
      id:
        type: integer
      level:
        type: integer
      max_distance:
        type: number
      total_kills:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  game_eating_pizza_internal_api_dto.WeaponResponse:
    properties:
      attack_power:
        type: integer
      attack_speed:
        type: number
      created_at:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      player_id:
        type: integer
      rarity:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  internal_api_handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_api_handlers.RegisterRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 횡스크롤 방치형 게임 백엔드 API 서버
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Game Eating Pizza API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 플레이어 로그인을 처리하고 JWT 토큰을 반환합니다
      parameters:
      - description: 로그인 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
      summary: 로그인
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: JWT 토큰을 갱신합니다
      produces:
      - application/json
      responses:
        "200":
          description: 토큰 갱신 성공
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
      summary: 토큰 갱신
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 새로운 플레이어를 등록합니다
      parameters:
      - description: 회원가입 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 회원가입 성공
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      summary: 회원가입
      tags:
      - auth
  /dungeons/{id}:
    get:
      consumes:
      - application/json
      description: 특정 던전의 상세 정보를 조회합니다
      parameters:
      - description: 던전 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 던전 정보
          schema:
            $ref: '#/definitions/game_eating_pizza_internal_api_dto.DungeonResponse'
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 던전을 찾을 수 없음
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 던전 상세 조회
      tags:
      - dungeons
  /dungeons/{id}/clear:
    post:
      consumes:
      - application/json
      description: 던전을 클리어하고 보상을 획득합니다
      parameters:
      - description: 던전 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 클리어 성공
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 던전 클리어
      tags:
      - dungeons
  /dungeons/{id}/enter:
    post:
      consumes:
      - application/json
      description: 던전에 입장합니다
      parameters:
      - description: 던전 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 입장 성공
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 던전 입장
      tags:
      - dungeons
  /dungeons/active:
    get:
      consumes:
      - application/json
      description: 활성화된 던전 목록만 조회합니다
      produces:
      - application/json
      responses:
        "200":
          description: 던전 목록
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/game_eating_pizza_internal_api_dto.DungeonResponse'
              type: array
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 활성 던전 목록 조회
      tags:
      - dungeons
  /dungeons/all:
    get:
      consumes:
      - application/json
      description: 전체 던전 목록을 조회합니다 (활성/비활성 포함)
      produces:
      - application/json
      responses:
        "200":
          description: 던전 목록
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/game_eating_pizza_internal_api_dto.DungeonResponse'
              type: array
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 전체 던전 목록 조회
      tags:
      - dungeons
  /players/leaderboard:
    get:
      consumes:
      - application/json
      description: 레벨이 높은 상위 플레이어 목록을 조회합니다
      parameters:
      - description: '조회할 플레이어 수 (기본값: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 리더보드
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 리더보드 조회
      tags:
      - players
  /players/me:
    get:
      consumes:
      - application/json
      description: 현재 로그인한 플레이어의 정보를 조회합니다
      produces:
      - application/json
      responses:
        "200":
          description: 플레이어 정보
          schema:
            $ref: '#/definitions/game_eating_pizza_internal_api_dto.PlayerResponse'
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 플레이어를 찾을 수 없음
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 내 정보 조회
      tags:
      - players
    put:
      consumes:
      - application/json
      description: 현재 로그인한 플레이어의 정보를 수정합니다
      produces:
      - application/json
      responses:
        "200":
          description: 수정 성공
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 내 정보 수정
      tags:
      - players
  /weapons:
    get:
      consumes:
      - application/json
      description: 현재 로그인한 플레이어의 무기 목록을 조회합니다
      produces:
      - application/json
      responses:
        "200":
          description: 무기 목록
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 무기 목록 조회
      tags:
      - weapons
    post:
      consumes:
      - application/json
      description: 새로운 무기를 생성합니다
      produces:
      - application/json
      responses:
        "201":
          description: 무기 생성 성공
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 무기 생성
      tags:
      - weapons
  /weapons/{id}/equip:
    put:
      consumes:
      - application/json
      description: 무기를 장착하여 현재 무기로 설정합니다
      parameters:
      - description: 무기 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 장착 성공
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 무기 장착
      tags:
      - weapons
  /weapons/{id}/upgrade:
    put:
      consumes:
      - application/json
      description: 무기를 강화하여 공격력을 증가시킵니다
      parameters:
      - description: 무기 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 강화된 무기 정보
          schema:
            $ref: '#/definitions/game_eating_pizza_internal_api_dto.WeaponResponse'
        "400":
          description: 잘못된 요청
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 서버 오류
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 무기 강화
      tags:
      - weapons
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"